move service config files into single directory
