+++ components/Feed/Comments/CommentHeader.tsx
@@ -10,10 +10,10 @@ import {
 import { useRouter } from 'next/router'
 import React, { FC } from 'react'
 
-import { db } from '../../../firebase'
 import { useProfileData } from '../../../hooks/useProfileData'
 import { getAuthorName, getProfilePic } from '../../../lib/profileHelper'
 import { deleteMedia } from '../../../lib/storageHelper'
+import { db } from '../../../services/firebase'
 import { postCardClass } from '../../../styles/feed'
 import { authorLabel } from '../../../utils/constants/global'
 import { staticPostData } from '../../../utils/types/params'
