+++ components/Feed/Post/PostHeader.tsx
@@ -10,11 +10,11 @@ import {
 import { useRouter } from 'next/router'
 import React, { FC } from 'react'
 
-import { db } from '../../../firebase'
 import useMediaQuery from '../../../hooks/useMediaQuery'
 import { useProfileData } from '../../../hooks/useProfileData'
 import { getAuthorName, getProfilePic } from '../../../lib/profileHelper'
 import { deleteMedia } from '../../../lib/storageHelper'
+import { db } from '../../../services/firebase'
 import { postCardClass } from '../../../styles/feed'
 import { FirebaseProfile } from '../../../utils/types/firebase'
 import { staticPostData } from '../../../utils/types/params'
