+++ hooks/useUserHasLiked.ts
@@ -1,7 +1,7 @@
 import { doc, onSnapshot } from 'firebase/firestore'
 import { useEffect, useState } from 'react'
 
-import { db } from '../firebase'
+import { db } from '../services/firebase'
 
 /**
  *
