+++ lib/commentsHelper.ts
@@ -9,7 +9,7 @@ import {
     where,
 } from 'firebase/firestore'
 
-import { db } from '../firebase'
+import { db } from '../services/firebase'
 
 /**
  *
