+++ lib/db.js
@@ -1,7 +1,7 @@
 import firebase from 'firebase/compat/app'
 import { getFunctions, httpsCallable } from 'firebase/functions'
 
-import { db } from '../firebase'
+import { db } from '../services/firebase'
 
 export async function createUser(uid, data) {
     const docRef = db.collection('users').doc(uid)
