+++ lib/postsHelper.ts
@@ -1,6 +1,6 @@
 import { doc, getDoc, onSnapshot } from 'firebase/firestore'
 
-import { db } from '../firebase'
+import { db } from '../services/firebase'
 
 /**
  *
