+++ pages/api/feeds.ts
@@ -7,7 +7,7 @@ import {
 } from 'firebase/firestore'
 import { NextApiRequest, NextApiResponse } from 'next'
 
-import { db } from '../../firebase'
+import { db } from '../../services/firebase'
 
 async function handleGet(res: NextApiResponse, _index: string, _order: string) {
     try {
